pyproject.toml
[project]
name = "nuu-dictate"
version = "0.1.0"
description = "Voice-to-text overlay MVP with hotkey activation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pynput>=1.7.6",
    "pyaudio>=0.2.11",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pytest>=8.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
nuu-dictate = "nuu_dictate.cli:main"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=nuu_dictate"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"
